FPAD-8280: Move InterventionState and InterventionName to constants
This allows them to be used in scripts that don't have access to AppConfigService
